[huge] fix (most) of rubocop offenses
This is a big commit, it was inevitable given that rubocop was
introduced after most of the code was written. No biggie.

Bunch of refactors, really helped to find code smells and separate
concerns of the objects, as well as to document the code (now at least
points what the class is meant to do).

10/10 will use rubocop again, it was great!